+++ ile_de_france/src/main/java/org/eqasim/ile_de_france/policies/routing/RoutingPenalty.java
@@ -1,0 +1,8 @@
+package org.eqasim.ile_de_france.policies.routing;
+
+import org.matsim.api.core.v01.network.Link;
+import org.matsim.api.core.v01.population.Person;
+
+public interface RoutingPenalty {
+	double getLinkPenalty(Link link, Person person, double time, double baseDisutility);
+}