+++ CHANGELOG.md
@@ -6,6 +6,7 @@ included in the (note yet determined) next version number.
 
 **Development version**
 
+- add configurable policies for IDF
 - Introduce `travelTimeRecordingInterval` config option that decouples travel time writing from general analysis
 - Add eqasim_activities.csv for analysis
 - The cutters now take a GeoPackage file as an alterative to a ShapeFile